VeinChain/backend +smart contracts /requirements.txt
Flask
web3
pycryptodome
flasgger
numpy
pillow
scikit-image
torch
torchvision
gunicorn
flask-cors
os
opencv-python
python-dotenv